Module reth_dns_discovery::tree

Support for the EIP-1459 DNS Record Structure

The nodes in a list are encoded as a merkle tree for distribution via the DNS protocol. Entries of the merkle tree are contained in DNS TXT records. The root of the tree is a TXT record with the following content:

enrtree-root:v1 e=<enr-root> l=<link-root> seq=<sequence-number> sig=<signature>

where

enr-root and link-root refer to the root hashes of subtrees containing nodes and links to subtrees. sequence-number is the tree’s update sequence number, a decimal integer. signature is a 65-byte secp256k1 EC signature over the keccak256 hash of the record content, excluding the sig= part, encoded as URL-safe base64 (RFC-4648).

Structs

BranchEntry

A branch entry with base32 hashes

LinkEntry

A link entry

NodeEntry

The actual Enr entry.

TreeRootEntry

Represents an enr-root hash of subtrees containing nodes and links.

Enums

DnsEntry

Represents all variants